requirements.txt
requests~=2.32.3
Flask~=3.1.0
python-dotenv~=1.0.1